+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- [fj-doc-playground-quarkus] base image registry.access.redhat.com/ubi9/openjdk-21:1.20-2.1729773471
+
 ## [8.10.4] - 2024-10-27
 
 ### Added