+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [8.10.5] - 2024-11-01
+
 ### Changed
 
 - [fj-doc-freemarker] generate stub and config conversion now supports registerById and allowDuplicatedId attributes